supporting PRGMA trusted_schema=0 and SQLITE_INNOCUOUS on functions and VTables
FossilOrigin-Name: 023f3323443fffa129c6438c2a42910de6811659

+++ test/check_add_tile_triggers_bad_table_name.c
@@ -76,6 +76,23 @@ main (int argc UNUSED, char *argv[]UNUSED)
 	  return -1;
       }
 
+    ret = sqlite3_exec (db_handle, "PRAGMA trusted_schema=0", NULL, NULL, &err_msg);
+    if (ret != SQLITE_OK)
+      {
+	  fprintf (stderr, "PRAGMA trusted_schema=0 error: %s\n", err_msg);
+	  sqlite3_free (err_msg);
+	  sqlite3_close (db_handle);
+	  return -1;
+      }
+    ret = sqlite3_exec (db_handle, "PRAGMA foreign_keys=1", NULL, NULL, &err_msg);
+    if (ret != SQLITE_OK)
+      {
+	  fprintf (stderr, "PRAGMA foreign_keys=1 error: %s\n", err_msg);
+	  sqlite3_free (err_msg);
+	  sqlite3_close (db_handle);
+	  return -1;
+      }
+
     /* create a minimal raster_format_metadata table (not spec compliant) */
     ret =
 	sqlite3_exec (db_handle, "DROP TABLE IF EXISTS raster_format_metadata",
